+++ packages/types/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @lezzform/types
 
+## 0.0.11
+
+### Patch Changes
+
+- custom messages support for all rules
+
+## 0.0.10
+
+### Patch Changes
+
+- change message to messages
+
+## 0.0.9
+
+### Patch Changes
+
+- add custom message types to all rules
+
 ## 0.0.8
 
 ### Patch Changes